final_pjt_back/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.7.1
Django==3.2.3
django-appconf==1.0.4
django-bootstrap-pagination==1.7.1
django-bootstrap-v5==1.0.1
django-cors-headers==3.7.0
django-extensions==3.1.1
django-imagekit==4.0.2
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
entrypoints==0.3
Faker==8.2.0
idna==2.10
importlib-metadata==2.1.1
inflection==0.5.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
pilkit==2.0
Pillow==8.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.18
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
Send2Trash==1.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
tmdb==0.0.0
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.1
